[ALPS03396073] memcfg: fix coverity issue
[Detail]
Cast memtotal to unsigned long long to avoid overflow.

MTK-Commit-Id: d7fa2315db863c76174e75b877d68fea4601080e

CR-Id: ALPS03396073
Feature: Memory

+++ drivers/misc/mediatek/mem/mtk_memcfg_reserve_info.c
@@ -205,7 +205,7 @@ static unsigned long long get_memtotal(void)
 	unsigned long long memtotal;
 
 	si_meminfo(&i);
-	memtotal = i.totalram << PAGE_SHIFT;
+	memtotal = (unsigned long long)i.totalram << PAGE_SHIFT;
 
 	return memtotal;
 }
